Split validation tests into A100 and H100 devices and update data handling
