+++ validation_scripts/imec_data/A100_inf.csv
@@ -1,0 +1,12 @@
+﻿device,model,TP,actual
+A100,Llama 2-7B,1,2.19
+A100,Llama 2-7B,2,1.544
+A100,Llama 2-7B,4,1.28
+A100,Llama 2-7B,8,1.187
+A100,Llama 2-13B,1,3.884
+A100,Llama 2-13B,2,2.499
+A100,Llama 2-13B,4,1.894
+A100,Llama 2-13B,8,1.693
+A100,Llama 2-70B,2,10.5
+A100,Llama 2-70B,4,6.403
+A100,Llama 2-70B,8,4.735
