--- validation_scripts/nvidia_inf.csv
@@ -1,23 +1,0 @@
-﻿device,model,TP,actual,seconds
-A100,Llama 2-7B,1,2.19,2.36
-A100,Llama 2-7B,2,1.544,1.82
-A100,Llama 2-7B,4,1.28,1.77
-A100,Llama 2-7B,8,1.187,2.16
-A100,Llama 2-13B,1,3.884,3.84
-A100,Llama 2-13B,2,2.499,2.75
-A100,Llama 2-13B,4,1.894,2.46
-A100,Llama 2-13B,8,1.693,2.85
-A100,Llama 2-70B,2,10.5,9.55
-A100,Llama 2-70B,4,6.403,6.98
-A100,Llama 2-70B,8,4.735,6.74
-H100,Llama 2-7B,1,1.44,1.82
-H100,Llama 2-7B,2,1.143,1.55
-H100,Llama 2-7B,4,0.924,1.62
-H100,Llama 2-7B,8,0.828,2.06
-H100,Llama 2-13B,1,2.396,2.75
-H100,Llama 2-13B,2,1.717,2.2
-H100,Llama 2-13B,4,1.431,2.17
-H100,Llama 2-13B,8,1.201,2.67
-H100,Llama 2-70B,2,6.267,6.77
-H100,Llama 2-70B,4,4.1226,5.56
-H100,Llama 2-70B,8,3.202,6
